upstream joshuaduffy/master
