+++ MANIFEST.in
@@ -1,1 +1,11 @@
-recursive-include dota2api/ref *
+include *.rst
+include *.py
+include *.pip
+include LICENSE
+
+recursive-include dota2api/ref *.json
+recursive-include dota2api *.py
+prune build
+prune dist
+prune docs
+prune *.egg-info
