+++ README.rst
@@ -8,7 +8,7 @@ dota2api: wrapper and parser
 
 Wrapper and parser in Python created for interacting and getting data easily from Valve's Dota 2 API.
 
-This libary parses some ID's into the dictionary keys like ``hero_name`` and so on. See ``src.parse`` for details.
+This library parses some ID's into the dictionary keys like ``hero_name`` and so on. See ``src.parse`` for details.
 
 This also comes with a growing set of tests and some documentation for the API itself.
  
@@ -16,8 +16,8 @@ Look how easy it is...
 
 .. code-block:: python
 
-    >>> from dota2api import wrapper
-    >>> api = wrapper.Initialise("API_KEY")
+    >>> import dota2api
+    >>> api = dota2api.Initialise("API_KEY")
     >>> hist = api.get_match_history(account_id=41231571)
     >>> match = api.get_match_details(match_id=1000193456).dict
     >>> match['radiant_win']
@@ -28,6 +28,31 @@ You can even store your API key as an environment variable instead of passing it
 .. code-block:: bash
 
     $  export D2_API_KEY=83247983248793298732
+
+
+Install
+-------
+
+Install via pip...
+
+.. code-block:: bash
+
+    $ pip install dota2api
+
+
+Or the old fashioned way...
+
+.. code-block:: bash
+
+    $ git clone https://github.com/joshuaduffy/dota2api.git
+    $ cd dota2api
+    $ python setup.py install
+
+
+Documentation
+-------------
+Documentation is available at http://dota2api.readthedocs.org/
+
 
 Supported API calls
 -------------------
@@ -41,21 +66,8 @@ Supported API calls
 - get_tournament_prize_pool
 - get_game_items
 
+
 Unsupported
 -----------
 - EconomySchema
 - GetMatchHistoryBySequenceNum
-
-
-Documentation
--------------
-Documentation is available at http://dota2api.readthedocs.org/
-
-Install
--------
-
-.. code-block:: bash
-
-    $ git clone https://github.com/joshuaduffy/dota2api.git
-    $ cd dota2api
-    $ python setup.py install
