+++ docs/source/examples.txt
@@ -1,0 +1,5 @@
+#################
+Examples
+#################
+
+TBC. But a copy of some of the responses will be here.
