+++ docs/source/index.txt
@@ -18,6 +18,8 @@ Contents
 
    installation
    tutorial
+   examples
+   parser
    reference
    contribute
