+++ docs/source/parser.txt
@@ -1,0 +1,5 @@
+#################
+Parser
+#################
+
+TBC. But a description of the parsing and how to get the values is here
