+++ setup.py
@@ -11,16 +11,17 @@ setup(
     version=dota2api.__version__,
     author=dota2api.__author__,
     author_email="mail@joshuaduffy.org",
+    url="https://github.com/joshuaduffy/dota2api",
     description=dota2api.__doc__,
     license=dota2api.__licence__,
     keywords="dota2 dota api dota2api parser",
     packages=['dota2api', 'dota2api.src'],
-    data_files=[('dota2api/ref', ['dota2api/ref/abilities.json',
-                                  'dota2api/ref/heroes.json',
-                                  'dota2api/ref/items.json',
-                                  'dota2api/ref/lobbies.json',
-                                  'dota2api/ref/modes.json',
-                                  'dota2api/ref/regions.json'])],
+    data_files=[('Lib/site-packages/dota2api/ref', ['dota2api/ref/abilities.json',
+                                                    'dota2api/ref/heroes.json',
+                                                    'dota2api/ref/items.json',
+                                                    'dota2api/ref/lobbies.json',
+                                                    'dota2api/ref/modes.json',
+                                                    'dota2api/ref/regions.json'])],
     install_requires=['requests'],
     classifiers=[
         "Development Status :: 4 - Beta",
